requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.0
pytest-django==4.8.0
sqlparse==0.5.0
